requirements.txt
amqp==5.0.2
asgiref==3.3.1
astroid==2.4.2
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.3
django-allauth==0.43.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
isort==5.6.4
kombu==5.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==8.0.1
prompt-toolkit==3.0.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
